APT: Move the shared font loading and relocation functions to their own subdirectory services/apt/bcfnt.
